producer/requirements.txt
pandas
openpyxl
kafka-python